Refactor for "Sum And Divide" Challenge Added "Generate Pattern" Challenge
