+++ main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"LeetCode/generate_pattern"
+	"LeetCode/sum_and_divide"
+)
+
+func main() {
+	sum_and_divide.SumAndDivide(4, 10)
+	generate_pattern.GeneratePattern(10)
+}
